moved tests to more appropriate places
