+++ tests/transformations/loop_to_map_test.py
@@ -12,6 +12,8 @@ from dace.sdfg import nodes
 from dace.sdfg.state import LoopRegion
 from dace.transformation.interstate import LoopToMap, StateFusion
 from dace.transformation.interstate.loop_lifting import LoopLifting
+from dace.transformation.interstate import LoopNormalize
+from dace.transformation.passes import SymbolPropagation
 
 
 def make_sdfg(with_wcr, map_in_guard, reverse_loop, use_variable, assign_after, log_path):
@@ -782,6 +784,100 @@ def test_self_loop_to_map():
     sdfg(A=a, N=20)
     assert np.allclose(a, ref)
 
+def _count_loops(sdfg):
+    loops = 0
+    for node, state in sdfg.all_nodes_recursive():
+        if isinstance(node, LoopRegion):
+            loops += 1
+    return loops
+
+
+def test_pipeline():
+    """
+    Tests that the Loop2Map pipeline (LoopNormalize -> SymbolPropagation) enables additional Loop2Map transformations.
+    """
+    sdfg = dace.SDFG("l2m_pipeline")
+    sdfg.add_array("A", [64], dace.float32)
+    sdfg.add_symbol("LB", dace.int32)
+    sdfg.add_symbol("UB", dace.int32)
+    sdfg.add_symbol("idx", dace.int32)
+
+    loop = LoopRegion("loop", "i < UB", "i", "i = LB", "i = i + 1")
+    sdfg.add_node(loop)
+    s = loop.add_state(is_start_block=True)
+    e = loop.add_state()
+    loop.add_edge(s, e, dace.InterstateEdge(assignments={"idx": "i - LB"}))
+    task = e.add_tasklet("init", {}, {"out"}, "out = 0")
+    access = e.add_access("A")
+    e.add_edge(task, "out", access, None, dace.Memlet("A[idx]"))
+
+    sdfg.validate()
+
+    # Count loops before transformation
+    assert _count_loops(sdfg) == 1
+
+    # Apply Loop2Map directly
+    sdfg.apply_transformations_repeated(LoopToMap)
+
+    # Should not have changed
+    assert _count_loops(sdfg) == 1
+
+    # Apply LoopNormalize and SymbolPropagation
+    sdfg.apply_transformations_repeated(LoopNormalize)
+    SymbolPropagation().apply_pass(sdfg, {})
+    sdfg.apply_transformations_repeated(LoopToMap)
+
+    # Should have transformed the loop
+    assert _count_loops(sdfg) == 0
+
+    # Validate correctness
+    A = dace.ndarray([64], dtype=dace.float32)
+    A[:] = np.random.rand(64).astype(dace.float32.type)
+    sdfg(A=A, LB=0, UB=64)
+
+    assert np.allclose(A[:], 0)
+
+
+def test_views():
+    """
+    Tests that the Loop2Map works correctly with views.
+    """
+    sdfg = dace.SDFG("l2m_pipeline")
+    sdfg.add_array("A", [64], dace.float32)
+    sdfg.add_view("A_view", [1], dace.float32)
+    sdfg.add_array("B", [64], dace.float32)
+
+    loop = LoopRegion("loop", "i < 64", "i", "i = 0", "i = i + 1")
+    sdfg.add_node(loop)
+    s = loop.add_state(is_start_block=True)
+    access_A = s.add_access("A")
+    access_view = s.add_access("A_view")
+    access_view.add_in_connector("views")
+    access_B = s.add_access("B")
+    tasklet = s.add_tasklet("copy", {"v"}, {"out"}, "out = v")
+    s.add_edge(access_A, None, access_view, "views", dace.Memlet("A[i]"))
+    s.add_edge(access_view, None, tasklet, "v", dace.Memlet("A_view[i]"))
+    s.add_edge(tasklet, "out", access_B, None, dace.Memlet("B[i]"))
+
+    sdfg.validate()
+
+    # Try to apply Loop2Map directly
+    try:
+        sdfg.apply_transformations_repeated(LoopToMap)
+        sdfg.validate()
+    except Exception as e:
+        assert False, f"LoopToMap failed: {e}"
+
+    # Validate correctness
+    A = dace.ndarray([64], dtype=dace.float32)
+    A[:] = np.random.rand(64).astype(dace.float32.type)
+    B = dace.ndarray([64], dtype=dace.float32)
+    B[:] = np.random.rand(64).astype(dace.float32.type)
+    sdfg(A=A, B=B)
+
+    for i in range(32):
+        assert np.isclose(A[i*2], B[i])
+
 
 if __name__ == "__main__":
 
@@ -822,3 +918,5 @@ if __name__ == "__main__":
     test_rotated_loop_to_map(False)
     test_rotated_loop_to_map(True)
     test_self_loop_to_map()
+    test_pipeline()
+    test_views()
